Add missing call to SPI.usingInterrupt()
The RFM69 driver accesses SPI in interrupt context.  This can (and does)
cause a lockup if another SPI device is being accessed in non-interrupt
context at the moment that the RFM69 interrupt occurs.

I had this problem on a board with both an Ethernet2 device and an RFM69
device.  Occasionally, the SPI interface would lock up hard, crashing
the application.

The fix is simple.  One must call SPI.usingInterrupt() and pass in the
interrupt number during setup.  The SPI library will record that fact,
and will automatically disable the interrupt whenever the SPI interface
is busy.

More information is available here:

https://github.com/arduino/Arduino/pull/2381

Please note that this is a race condition, and is somewhat hard to
trigger.  You need a board with several active SPI devices, and the
timing has to be such that interrupts happen in the middle of other SPI
transactions.